Neaten up

+++ src/hmi/gui/screens/dashboard/controller_pane/controllers/controller.h
@@ -11,9 +11,9 @@ void controller_set_label(lv_obj_t* controller, const char* label);
 
 void controller_set_options(lv_obj_t* controller, const char* options);
 
-void controller_select_option(lv_obj_t* controller, uint16_t optionIndex);
+void controller_select_option(lv_obj_t* controller, uint8_t optionIndex);
 
-typedef void(*option_change_handler)(uint16_t optionIndex);
+typedef void(*option_change_handler)(uint8_t optionIndex);
 void controller_handle_option_change(lv_obj_t* controller, option_change_handler callback);
 
 #endif // !CONTROLLER_H
